codec2: fix checking flush state in onQueueWork
In CCodec aspect, flush is complete after flush signal is finished and would
not be blocked. So actually CCodec will not know the internal state of the
component, and it is valid for CCodec to start queuing works then.

The component should internally block the work queue until flush is complete
(just like we did for draining).

Bug: 80452412
Test: CTsMediaTestCases android.media.cts.DecoderTest#testCodecResetsH264WithSurface

Change-Id: I01b2af1ff4a94eff2ad3cdaebbe9c44709a04d8c