codec2: put abandoned works to another vector apart from mPendingWorks
Flush signal would be succeeded by draining, for this case in onDrainDone() we
should also expect state in FLUSHING and still wait for onFlushDone().

In order to handle reportEOSWork() by draining for this case, we should not mix
up pending works (waiting for VDA processed) and abandoned works. A vector
mAbandonedWorks is introduced to temporarily place abandoned works, and passed
to listener for abandoning in reportAbandonedWorks() call.

Bug: 80452475
Test: CtsMediaTestCases android.media.cts.MediaPlayerTest#testLocalVideo_MP4_H264_480x360_1000kbps_25fps_AAC_Stereo_128kbps_44110Hz
Change-Id: I38ed72222a2276f85485e57409adc88e53fe416e