codec2: C2VDAComponent: use bufferpool-backed block pool
When there is no surface (byte-buffer mode), C2VDAComponent will use
C2VdaPooledBlockPool for allocating and recycling graphic buffers.

The hack of MockBufferInClient is removed and the dequeue flow of
C2VdaPooledBlockPool is aligned to C2VdaBqBlockPool.

Component is not always holding the reference of C2GraphicBlock now, it is
obtained from block pool and passed to the client after decoding.

Bug: 79239042
Test: CtsMediaTestCases on kevin-arcnext with Codec2.0 enabled
Test: android.media.cts.EncodeDecodeTest#testEncodeDecodeVideoFromBufferToBuffer720p
Test: android.media.cts.EncodeDecodeTest#testEncodeDecodeVideoFromBufferToBufferQCIF
Test: android.media.cts.EncodeDecodeTest#testEncodeDecodeVideoFromBufferToBufferQVGA
Test: android.media.cts.EncodeDecodeTest#testVP8EncodeDecodeVideoFromBufferToBuffer720p
Test: android.media.cts.EncodeDecodeTest#testVP8EncodeDecodeVideoFromBufferToBufferQCIF
Test: android.media.cts.EncodeDecodeTest#testVP8EncodeDecodeVideoFromBufferToBufferQVGA
Change-Id: I31d2caf87ac3b60c46dc642c18cd0d254f628fb2