Use C2VDAPixelForamt to resolve pixel format
A function to get supported format from crcb and semiplanar is identical in Android path and ARC++
path. So they can be unified.
The other function is to return pixel format on a platform with zero knowledge. A small
C2AllocationGralloc is allocated inside of the function, and detects its pixel format.
Because C2ArcProtectedGraphicAllocator needs to call these function, the code is located in
device/google/cheets2/.

Bug: 73870167
Test: Play secure videos with ExoPlayer.app

Change-Id: I34975f869b465e347b9ef1521d12998539be5658