C2VDAComponent: do not rely on monotonic increasing output timestamp
Unfortuneately the previous assumption of getting monotonic increasing
timestamp as display order of output buffers is not always right. In
testH264_adaptiveSkipBack, the timestamp goes back and forth.

For such cases we cannot rely on timestamp for checking work done anymore, we
cannot help but rely on input flags for recognizing CSD buffer works, which
doesn't have a correlated output buffer; otherwise, work is done only if both
input and output buffer are returned from VDA.

Bug: 80402878
Test: android.media.cts.AdaptivePlaybackTest#testH264_adaptiveSkipBack for
codec2.0 on eve-arcnext

Change-Id: Ib3af3ca0ee90925187662af4ae0eb4d628ad21a6