C2VDAComponent: Get fds of all the planes of output buffer
C2VDAComponent only gets the fd of the first plane of output buffer. This works because all the
planes are stored in the same buffer. The fds of second or later planes refers the same buffer.
This generalizes the C2VDAComponent code by getting fds of all the planes and sends to C2VDAAdaptor.

Bug: 131122595
Test: Play movie with VLC app
Change-Id: Ibfe843f20763546ab73caf9061ed657182ae9299
(cherry picked from commit fd22b9d58807655615a86faef11e9641787e1ba2)