C2VEAComponent: match pending work with returned output buffer by timestamp
For C2Work, the output worklet should be assigned with the output bitstream
which is encoded from the corresponding input frame. In VEA's aspect, the
callback of bitstream ready is paired with timestamp, which is obtained from
the corresponding input frame. That is, the component should use timestamp to
match the output bitstream with the pending work.

Bug: 73059339
Test: arcvideoencoder native e2e test on Soraka and Kevin

Change-Id: I235a3d8aa2f1e33de762a25a1e8ce2414a18e629
(cherry picked from commit c541e2c0d99424ced9294b4f7a95b1753e1b7e79)