C2VEAComponent: do not assert on checking tasks under running state
In the past we check most of the function tasks to be executed while component
state is running (not UNINITIALIZED) and assert if violated. However, in some
cases we might not guarantee there is no remaining task after onStop() is
executed, which will change the state to UNINITIALIZED immediately.

Instead of assertion, we modified to just return if tasks are executed while
component state is UNINITIALIZED. It makes sense because the component is
already released everything. And it could also avoid unnecessary process
crashes.

Bug: 73059339
Test: android.media.cts.MediaCodecTest#testAbruptStop passed and no crash
Change-Id: Ib5b68ed51f3ba137b981edc24519dd4ba7144532
(cherry picked from commit 23cccc9f965581241573bad6a9f840a8363bb0d7)