C2VEAComponent: apply buffer index as identification on BitstreamBufferReady callback
In VEA API, UseBitstreamBuffer() doesn't require to correlate to Encode(). That
is, in C2VEAComponent's aspect, output buffers are not guaranteed to be FIFO.

Instead, we should apply a unique buffer index for each output buffer and
assign it to the corresponding BitstreamBufferReady callback. It could help
C2VEAComponent to identify which output buffer is returned.

Bug: 73059339
Test: pass arcvideoencoder native e2e test on Kevin.
Test: pass CTS android.media.cts.EncodeDecodeTest on Kevin.
Change-Id: I0c7a7768f9266c3c908dfe4506cbffa47a3459a4
(cherry picked from commit 9bb2cd5745bee36c877b67bc6ad16656cf6eb574)