v4l2_codec2: Implement Codec2 V4L2 encode component.
This CL introduces the first version of the new codec2 V4L2 encoder.
The V4L2EncodeComponent currently only supports H264, support for
additional formats will introduced later.

The encode component is developed and tested on ARC++ and the initial
version is not functional yet on ARCVM. The changes required to fully
enable the encoder on ARCVM (such as support for DMABufs) will be
introduced in follow-up CLs.

Currently the component makes heavy use of Chromium helpers as it's
based on the Chromium V4L2VideoEncodeAccelerator, but in the future
the amount of dependencies will be reduced.

The new V4L2 encoder is currently not enabled by default, which requires
additional changes in media_codecs_c2.xml.

Bug: 143333813
Test: mmm external/v4l2_codec2
Change-Id: I59ba52e8f85aa40c8d23a451fb2842a7a421412a