v4l2_codec2: Update e2e encoder tests to try both old and new encoders.
ag/11749538 updated the encoder tests to use the new V4L2 video encoder.
But a single test apk is built of the rvc-arc-dev branch for both ARC++
and ARCVM, causing all ARC++ video encoder tests to start failing.

This CL fixes this by adding the names of both ARC++ and ARCVM video
encoders so the tests tries both.

Bug: 160755362
Test: mmm external/v4l2_codec2/tests/c2_e2e_test
Change-Id: Ib31c711d365e5e33a5a54ae6cd9d852431066310