Change getPoolIdFromGraphicBlock() signature
getPoolIdFromGraphicBlock() is used to get the buffer index from
C2GraphicBlock. Originally the function takes the pointer for returing
the buffer index, and returns the status code. But actually we only
care if the function is successfully or not.

This CL changes the function signature to return a std::optional<>
value to indicate if the function successfully gets the buffer index.

Bug: 160110407
Test: run e2e test
Change-Id: I88e54831b1964d20cd888e2b69a028e75a3e50cd