VideoFramePool: use exponential backoff for obtaining frames
The VideoFramePool needs to poll its backing block pool until a buffer
becomes available. The current strategy is to sleep and repeat a number
of times before giving up, but this carries the risk that some slower
platforms will fail to receive buffers in the alloted time and introduce
flakyness. Also the constant time between requests results in many
requests to the block pool when it is empty, a situation that typically
corresponds to the decoder having plenty of buffers at its disposal and
thus not being in a particular hurry for a new buffer.

Switch to an exponential backoff strategy for the block pool polling,
and remove the time boundary. The exponential backoff limits the number
of queries that are done to the block pool, and the absence of time
makes the whole process safer.  In case of a real error, the backing
block pool is supposed to return a proper error code that we will
propagate anyway.

Bug: 161323057
Test: arc.VideoDecodeAccel.h264_vm does not time out waiting for buffers
on Kukui.

Change-Id: Ic3b9915926680cfd263352a7922658ec9a60b5cf