v4l2_codec2: Remove code that prepends SPS/PPS to IDR frames.
This CL removes the code that prepends SPS/PPS to IDR frames, as this
is already handled by the Chrome video encoder.

Note: When we upstream this code we might want this functionality, but
as ag/12208262 makes changes that would require us to rewrite this
logic we can reintroduce it when we decide it's necessary.

Bug: 143333813
Test: tast run $HOST arc.VideoEncodeAccel.h264_192p_i420_vm
Change-Id: Ib0f7fac6ac23fca1b88c3a026989338ca2aaf95f