v4l2_codec2: Don't use BufferUsage::VIDEO_ENCODER for input buffers.
This CL removes the BufferUsage::VIDEO_ENCODER flag from input buffers.
Currently the format convertor is always enabled as the encoder on
Chrome side only supports the NV12 format. This means it's not useful
to add the VIDEO_ENCODER buffer flag to input buffers. Currently
zero-copy is not supported yet, so when using that flag an additional
buffer will be allocated on host side and a copy will be performed
between the guest and host buffer to keep them in sync. This is
wasteful as the buffer is only used on guest side by the format
convertor which converts and copies the buffer into another buffer.

Additionally this causes issues on hatch as the YV12 format is used
which is not supported by minigbm on hatch. (see b/162796596)

Bug: 162796596
Test: tast run $HOST arc.VideoEncodeAccel.h264_360p_i420_vm on
      hatch-arc-r and verified encoded video
Change-Id: Iffcaafbe12fa1f9af53ac7c4e88ee98ec40aedec