V4L2Decoder: streamon V4L2 output queue in Flush()
Originally we streamoff both input and output queue when Flush() is
called. We expect the V4L2 driver will send the resolution change
event after the playback resumes. Then the output queue will be
streamon again.

However, the driver might not send the event if the resumed playback
has the same resolution. In this case, the output queue won't be
streamon and the playback is stuck. This CL streamons the output queue
immediately after the output queue is streamoff.

Bug: 168557465
Test: Pass android.media.cts.AdaptivePlaybackTest#testVP8_eosFlushSeek
      with related CLs

Change-Id: I0e3f0bd6445150ed9df2a30aac8ff692ec626137